cloud_roles/uri_encode: move function to request_response_helpers
