+++ src/v/cloud_roles/request_response_helpers.cc
@@ -207,4 +207,33 @@ std::chrono::system_clock::time_point parse_timestamp(std::string_view sv) {
     return std::chrono::system_clock::from_time_t(timegm(&tm));
 }
 
+inline void append_hex_utf8(ss::sstring& result, char ch) {
+    bytes b = {static_cast<uint8_t>(ch)};
+    result.append("%", 1);
+    auto h = to_hex(b);
+    result.append(h.data(), h.size());
+}
+
+ss::sstring uri_encode(const ss::sstring& input, bool encode_slash) {
+    // The function defined here:
+    //     https://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-header-based-auth.html
+    ss::sstring result;
+    for (auto ch : input) {
+        if (
+          (ch >= 'A' && ch <= 'Z') || (ch >= 'a' && ch <= 'z')
+          || (ch >= '0' && ch <= '9') || (ch == '_') || (ch == '-')
+          || (ch == '~') || (ch == '.')) {
+            result.append(&ch, 1);
+        } else if (ch == '/') {
+            if (encode_slash) {
+                result.append("%2F", 3);
+            } else {
+                result.append(&ch, 1);
+            }
+        } else {
+            append_hex_utf8(result, ch);
+        }
+    }
+    return result;
+}
 } // namespace cloud_roles
